gst/qtdemux/qtdemux.c: Make timescale 32 bits again so we don't screw up the pts_offset calculations.
Original commit message from CVS:
* gst/qtdemux/qtdemux.c:
Make timescale 32 bits again so we don't screw up the pts_offset
calculations.

+++ gst/qtdemux/qtdemux.c
@@ -117,7 +117,7 @@ struct _QtDemuxStream
 
   /* duration/scale */
   guint64 duration;             /* in timescale */
-  guint64 timescale;
+  guint32 timescale;
 
   /* our samples */
   guint32 n_samples;
